finished nestjs project

+++ src/products/products.controller.ts
@@ -1,8 +1,11 @@
+import { UpdateProductDTO } from './dtos/update-product.dto';
 import { Controller, Get } from '@nestjs/common';
-import { Body, Delete, Param, Post } from '@nestjs/common/decorators';
+import { Body, Delete, Param, Post,  Put } from '@nestjs/common/decorators';
 import { Product } from 'src/db';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { ParseUUIDPipe } from '@nestjs/common/pipes';
+import { NotFoundException } from '@nestjs/common/exceptions';
 
 @Controller('products')
 export class ProductsController {
@@ -16,19 +19,36 @@ export class ProductsController {
   }
 
   @Get('/:id')
-  public getById(@Param('id') id: string) {
-    return this.productsService.getById(id)
+  getById(@Param('id', new ParseUUIDPipe()) id: string) {
+    const prod = this.productsService.getById(id);
+    if (!prod) throw new NotFoundException('Product not found');
+    return prod;
   }
-
+  
   @Delete('/:id')
-  public deleteById(@Param('id') id: string) {
+  deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
     this.productsService.deleteById(id);
-    return { success: true};
+    return { success: true };
   }
 
   @Post('/')
   public create(@Body() productData: CreateProductDTO) {
     return this.productsService.create(productData);
   }
+
+  @Put('/:id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
+    this.productsService.updateById(id, productData);
+    return { success: true };
+  }
+  
   
 }
